introduce extHostLabelService and mainThreadLabelService
fixes #75881
